@@ -1,39 +1,34 @@
 import 'dart:async' show StreamController;
 
-import 'package:flutter/material.dart' show Widget;
+import 'package:flutter/material.dart';
 import 'package:moliseis/config/env/env.dart';
-import 'package:moliseis/data/repositories/attraction/attraction_repository.dart';
-import 'package:moliseis/data/repositories/attraction/attraction_repository_local.dart';
-import 'package:moliseis/data/repositories/gallery/gallery_repository.dart';
-import 'package:moliseis/data/repositories/gallery/gallery_repository_local.dart';
+import 'package:moliseis/data/repositories/city/city_repository.dart';
+import 'package:moliseis/data/repositories/city/city_repository_local.dart';
+import 'package:moliseis/data/repositories/event/event_repository.dart';
+import 'package:moliseis/data/repositories/event/event_repository_local.dart';
+import 'package:moliseis/data/repositories/geo_map/geo_map_repository.dart';
+import 'package:moliseis/data/repositories/geo_map/geo_map_repository_remote.dart';
+import 'package:moliseis/data/repositories/media/media_repository.dart';
+import 'package:moliseis/data/repositories/media/media_repository_local.dart';
 import 'package:moliseis/data/repositories/place/place_repository.dart';
 import 'package:moliseis/data/repositories/place/place_repository_local.dart';
 import 'package:moliseis/data/repositories/search/search_repository.dart';
 import 'package:moliseis/data/repositories/search/search_repository_local.dart';
 import 'package:moliseis/data/repositories/settings/settings_repository.dart';
 import 'package:moliseis/data/repositories/settings/settings_repository_local.dart';
-import 'package:moliseis/data/repositories/story/paragraph_repository.dart';
-import 'package:moliseis/data/repositories/story/paragraph_repository_local.dart';
-import 'package:moliseis/data/repositories/story/story_repository.dart';
-import 'package:moliseis/data/repositories/story/story_repository_local.dart';
 import 'package:moliseis/data/repositories/suggestion/suggestion_repository.dart';
 import 'package:moliseis/data/repositories/suggestion/suggestion_repository_remote.dart';
-import 'package:moliseis/data/services/remote/cloudinary.dart';
-import 'package:moliseis/domain/models/attraction/attraction_supabase_table.dart';
-import 'package:moliseis/domain/models/molis_image/image_supabase_table.dart';
-import 'package:moliseis/domain/models/paragraph/paragraph_supabase_table.dart';
+import 'package:moliseis/data/services/remote/cloudinary_client.dart';
+import 'package:moliseis/data/services/remote/openstreetmap_client.dart';
+import 'package:moliseis/domain/models/city/city_supabase_table.dart';
+import 'package:moliseis/domain/models/event/event_supabase_table.dart';
+import 'package:moliseis/domain/models/media/media_supabase_table.dart';
 import 'package:moliseis/domain/models/place/place_supabase_table.dart';
-import 'package:moliseis/domain/models/story/story_supabase_table.dart';
 import 'package:moliseis/domain/models/suggestion/suggestion_supabase_table.dart';
-import 'package:moliseis/domain/use-cases/sync/sync_start_use_case.dart';
+import 'package:moliseis/domain/use-cases/favourite/favourite_get_ids_use_case.dart';
+import 'package:moliseis/domain/use-cases/sync/sync_repo_use_case.dart';
 import 'package:moliseis/main.dart';
-import 'package:moliseis/ui/categories/view_models/categories_view_model.dart';
-import 'package:moliseis/ui/explore/view_models/attraction_view_model.dart';
-import 'package:moliseis/ui/explore/view_models/place_view_model.dart';
 import 'package:moliseis/ui/favourite/view_models/favourite_view_model.dart';
-import 'package:moliseis/ui/gallery/view_models/gallery_view_model.dart';
-import 'package:moliseis/ui/geo_map/view_models/geo_map_view_model.dart';
-import 'package:moliseis/ui/search/view_models/search_view_model.dart';
 import 'package:moliseis/ui/settings/view_models/settings_view_model.dart';
 import 'package:moliseis/ui/settings/view_models/theme_view_model.dart';
 import 'package:moliseis/ui/sync/view_models/sync_view_model.dart';
@@ -47,45 +42,44 @@ List<SingleChildWidget> get providers {
     //#region Repositories (sorted by name ascending)
     Provider(
       create: (_) {
-        return AttractionRepositoryLocal(
-              supabaseI: Supabase.instance,
-              attractionSupabaseTable: AttractionSupabaseTable(),
-              objectBoxI: objectBox,
-            )
-            as AttractionRepository;
-      },
-      lazy: true,
-    ),
-    Provider(
-      create: (_) {
-        return GalleryRepositoryLocal(
-              supabaseI: Supabase.instance,
-              imageSupabaseTable: ImageSupabaseTable(),
-              objectBoxI: objectBox,
-            )
-            as GalleryRepository;
-      },
-      lazy: true,
-    ),
-    Provider(
-      create: (_) {
-        return ParagraphRepositoryLocal(
-              supabaseI: Supabase.instance,
-              supabaseTable: ParagraphSupabaseTable(),
-              objectBoxI: objectBox,
-            )
-            as ParagraphRepository;
-      },
-      lazy: true,
-    ),
-    Provider(
-      create: (_) {
         return PlaceRepositoryLocal(
               supabaseI: Supabase.instance,
-              placeSupabaseTable: PlaceSupabaseTable(),
+              supabaseTable: PlaceSupabaseTable(),
               objectBoxI: objectBox,
             )
             as PlaceRepository;
+      },
+      lazy: true,
+    ),
+    Provider(
+      create: (_) {
+        return EventRepositoryLocal(
+              supabaseI: Supabase.instance,
+              supabaseTable: EventSupabaseTable(),
+              objectBoxI: objectBox,
+            )
+            as EventRepository;
+      },
+    ),
+    Provider(
+      create: (_) {
+        return MediaRepositoryLocal(
+              supabaseI: Supabase.instance,
+              imageSupabaseTable: MediaSupabaseTable(),
+              objectBoxI: objectBox,
+            )
+            as MediaRepository;
+      },
+      lazy: true,
+    ),
+    Provider(
+      create: (_) {
+        return CityRepositoryLocal(
+              supabaseI: Supabase.instance,
+              placeSupabaseTable: CitySupabaseTable(),
+              objectBoxI: objectBox,
+            )
+            as CityRepository;
       },
       lazy: true,
     ),
@@ -99,17 +93,6 @@ List<SingleChildWidget> get providers {
       create: (context) {
         return SettingsRepositoryLocal(objectBoxI: objectBox)
             as SettingsRepository;
-      },
-      lazy: true,
-    ),
-    Provider(
-      create: (_) {
-        return StoryRepositoryLocal(
-              supabaseI: Supabase.instance,
-              supabaseTable: StorySupabaseTable(),
-              objectBoxI: objectBox,
-            )
-            as StoryRepository;
       },
       lazy: true,
     ),
@@ -129,19 +112,12 @@ List<SingleChildWidget> get providers {
             as SuggestionRepository;
       },
     ),
-    //#endregion
-
-    //#region UseCases (sorted by name)
-    Provider<SyncStartUseCase>(
-      create: (context) {
-        return SyncStartUseCase(
-          attractionRepository: context.read(),
-          galleryRepository: context.read(),
-          //paragraphRepository: context.read(),
-          placeRepository: context.read(),
-          settingsRepository: context.read(),
-          //storyRepository: context.read(),
-        );
+    Provider<GeoMapRepository>(
+      create: (_) {
+        return GeoMapRepositoryRemote(
+              openStreetMapClient: OpenStreetMapClient(),
+            )
+            as GeoMapRepository;
       },
       lazy: true,
     ),
@@ -154,66 +130,34 @@ List<SingleChildWidget> get providers {
       },
       lazy: true,
     ),
+    ChangeNotifierProvider<SyncViewModel>(
+      create: (context) {
+        final useCase = SynchronizeRepositoriesUseCase(
+          cityRepository: context.read(),
+          eventRepository: context.read(),
+          mediaRepository: context.read(),
+          placeRepository: context.read(),
+          settingsRepository: context.read(),
+        );
+
+        return SyncViewModel(syncRepoUseCase: useCase);
+      },
+      lazy: true,
+    ),
     ChangeNotifierProvider<SettingsViewModel>(
       create: (context) {
         return SettingsViewModel(settingsRepository: context.read());
       },
       lazy: true,
     ),
-    ChangeNotifierProvider<AttractionViewModel>(
-      create: (context) {
-        return AttractionViewModel(attractionRepository: context.read());
-      },
-      lazy: true,
-    ),
-    ChangeNotifierProvider<GalleryViewModel>(
-      create: (context) {
-        return GalleryViewModel(
-          attractionRepository: context.read(),
-          galleryRepository: context.read(),
-        );
-      },
-      lazy: true,
-    ),
-    ChangeNotifierProvider<GeoMapViewModel>(
-      create: (context) {
-        return GeoMapViewModel(attractionRepository: context.read());
-      },
-      lazy: true,
-    ),
-    Provider<PlaceViewModel>(
-      create: (context) {
-        return PlaceViewModel(placeRepository: context.read());
-      },
-      lazy: true,
-    ),
-    ChangeNotifierProvider<SearchViewModel>(
-      create: (context) {
-        return SearchViewModel(
-          attractionRepository: context.read(),
-          searchRepository: context.read(),
-        );
-      },
-      lazy: true,
-    ),
-    ChangeNotifierProvider<CategoriesViewModel>(
-      create: (context) {
-        return CategoriesViewModel(attractionRepository: context.read());
-      },
-      lazy: true,
-    ),
-    ChangeNotifierProvider<SyncViewModel>(
-      create: (context) {
-        return SyncViewModel(
-          syncStartUseCase: context.read(),
-          settingsRepository: context.read(),
-        );
-      },
-      lazy: true,
-    ),
     ChangeNotifierProvider<FavouriteViewModel>(
       create: (context) {
-        return FavouriteViewModel(attractionRepository: context.read());
+        return FavouriteViewModel(
+          favouriteGetIdsUseCase: FavouriteGetIdsUseCase(
+            eventRepository: context.read(),
+            placeRepository: context.read(),
+          ),
+        );
       },
     ),
     //#endregion
