@@ -1,27 +1,27 @@
-import 'package:moliseis/domain/models/attraction/attraction_sort.dart';
+import 'package:moliseis/domain/models/core/content_sort.dart';
 import 'package:moliseis/domain/models/settings/theme_brightness.dart';
 import 'package:moliseis/domain/models/settings/theme_type.dart';
+import 'package:moliseis/utils/result.dart';
 
 abstract class SettingsRepository {
   /// Returns the user selected theme type (system, app).
   ThemeType get themeType;
 
-  void setThemeType(ThemeType type);
+  Future<Result<void>> setThemeType(ThemeType type);
 
   /// Returns the user selected theme brightness (system, light, dark).
   ThemeBrightness get themeBrightness;
 
   void setThemeBrightness(ThemeBrightness brightness);
 
-  /// Returns the user selected attractions sort order (by name or by
-  /// creation date).
-  AttractionSort get attractionSort;
+  /// Returns the user selected content sort order.
+  ContentSort get contentSort;
 
-  void setAttractionSort(AttractionSort sort);
+  Future<Result<void>> setContentSort(ContentSort sort);
 
   /// Returns the last time the repositories have been successfully synchronized
   /// with the backend.
-  DateTime get modifiedAt;
+  DateTime? get modifiedAt;
 
   void setModifiedAt(DateTime dateTime);
 
@@ -29,5 +29,5 @@ abstract class SettingsRepository {
   /// Sentry service or not.
   bool get crashReporting;
 
-  void setCrashReporting(bool enable);
+  Future<Result<void>> setCrashReporting(bool enable);
 }
