@@ -1,34 +1,48 @@
 import 'dart:async' show Future;
 
+import 'package:logging/logging.dart';
 import 'package:moliseis/data/repositories/settings/settings_repository.dart';
 import 'package:moliseis/data/services/objectbox.dart';
-import 'package:moliseis/domain/models/attraction/attraction_sort.dart';
+import 'package:moliseis/domain/models/core/content_sort.dart';
 import 'package:moliseis/domain/models/settings/app_settings.dart';
 import 'package:moliseis/domain/models/settings/theme_brightness.dart';
 import 'package:moliseis/domain/models/settings/theme_type.dart';
 import 'package:moliseis/generated/objectbox.g.dart';
+import 'package:moliseis/utils/result.dart';
 
 class SettingsRepositoryLocal implements SettingsRepository {
+  final Box<AppSettings> _settingsBox;
+
   SettingsRepositoryLocal({required ObjectBox objectBoxI})
     : _settingsBox = objectBoxI.store.box<AppSettings>();
 
-  final Box<AppSettings> _settingsBox;
+  final _log = Logger('SettingsRepositoryLocal');
 
   /// Loads from the database (or creates) the [AppSettings] object.
   ///
   /// ObjectBox does not support the creation of a box with only one object
   /// inside. As such, the [AppSettings.id] is not managed by ObjectBox and
   /// is always set to constant [settingsId] during object instantiation.
-  AppSettings get _appSettings =>
-      _settingsBox.get(settingsId) ??
-      AppSettings(modifiedAt: DateTime.parse('1970-01-01T00:00:00Z'));
+  AppSettings get _appSettings => _settingsBox.get(settingsId) ?? AppSettings();
 
   @override
   ThemeType get themeType => _appSettings.type;
 
   @override
-  Future<void> setThemeType(ThemeType type) =>
-      _settingsBox.putAsync(_appSettings.copyWith(type: type));
+  Future<Result<void>> setThemeType(ThemeType type) async {
+    try {
+      await _settingsBox.putAsync(_appSettings.copyWith(type: type));
+
+      return const Result.success(null);
+    } on Exception catch (error, stackTrace) {
+      _log.severe(
+        'An exception occurred while setting theme type to $type.',
+        error,
+        stackTrace,
+      );
+      return Result.error(error);
+    }
+  }
 
   @override
   ThemeBrightness get themeBrightness => _appSettings.brightness;
@@ -38,14 +52,25 @@ class SettingsRepositoryLocal implements SettingsRepository {
       _settingsBox.putAsync(_appSettings.copyWith(brightness: brightness));
 
   @override
-  AttractionSort get attractionSort => _appSettings.attractionSort;
+  ContentSort get contentSort => _appSettings.contentSort;
 
   @override
-  Future<void> setAttractionSort(AttractionSort sort) =>
-      _settingsBox.putAsync(_appSettings.copyWith(attractionSort: sort));
+  Future<Result<void>> setContentSort(ContentSort sort) async {
+    try {
+      await _settingsBox.putAsync(_appSettings.copyWith(contentSort: sort));
+      return const Result.success(null);
+    } on Exception catch (error, stackTrace) {
+      _log.severe(
+        'An exception occurred while setting content sort to $sort.',
+        error,
+        stackTrace,
+      );
+      return Result.error(error);
+    }
+  }
 
   @override
-  DateTime get modifiedAt => _appSettings.modifiedAt;
+  DateTime? get modifiedAt => _appSettings.modifiedAt;
 
   @override
   void setModifiedAt(DateTime dateTime) =>
@@ -55,6 +80,19 @@ class SettingsRepositoryLocal implements SettingsRepository {
   bool get crashReporting => _appSettings.crashReporting;
 
   @override
-  Future<void> setCrashReporting(bool enable) =>
-      _settingsBox.putAsync(_appSettings.copyWith(crashReporting: enable));
+  Future<Result<void>> setCrashReporting(bool enable) async {
+    try {
+      await _settingsBox.putAsync(
+        _appSettings.copyWith(crashReporting: enable),
+      );
+      return const Result.success(null);
+    } on Exception catch (error, stackTrace) {
+      _log.severe(
+        'An exception occurred while setting crash reporting to $enable.',
+        error,
+        stackTrace,
+      );
+      return Result.error(error);
+    }
+  }
 }
