@@ -1,11 +1,6 @@
-import 'dart:io' show Platform;
-
-import 'package:device_info_plus/device_info_plus.dart';
 import 'package:flutter/foundation.dart';
 import 'package:flutter/material.dart';
-import 'package:flutter/services.dart';
 import 'package:flutter_localizations/flutter_localizations.dart';
-import 'package:logging/logging.dart';
 import 'package:moliseis/config/dependencies.dart';
 import 'package:moliseis/config/env/env.dart';
 import 'package:moliseis/data/services/objectbox.dart';
@@ -18,10 +13,6 @@ import 'package:responsive_framework/responsive_framework.dart';
 import 'package:sentry_flutter/sentry_flutter.dart';
 import 'package:sentry_logging/sentry_logging.dart';
 import 'package:supabase_flutter/supabase_flutter.dart';
-
-part '_main.dart';
-
-late final Logger log;
 
 /// The ObjectBox instance.
 late final ObjectBox objectBox;
@@ -37,15 +28,36 @@ void main() async {
 
   objectBox = await ObjectBox.create();
 
-  log = Logger('Molise Is');
-
   // Retrieves the app settings to check whether the user has given his consent
   // to report exceptions or not.
   final settings = objectBox.store.box<AppSettings>().get(settingsId);
 
   final reportExceptions = (settings?.crashReporting ?? false) && !kDebugMode;
 
-  await _initializeSentry(reportExceptions, const MoliseIsApp());
+  await _initializeSentry(enable: reportExceptions, app: const MoliseIsApp());
+}
+
+Future<void> _initializeSentry({
+  required bool enable,
+  required Widget app,
+}) async {
+  if (enable) {
+    await SentryFlutter.init((options) {
+      options.dsn = Env.sentryUrl;
+      options.addIntegration(LoggingIntegration());
+      // Set tracesSampleRate to 1.0 to capture 100% of transactions for
+      // tracing.
+      options.tracesSampleRate = 0.4;
+      // The sampling rate for profiling is relative to tracesSampleRate
+      // Setting to 1.0 will profile 100% of sampled transactions:
+      options.profilesSampleRate = 1.0;
+      // Session Replay setup.
+      options.replay.sessionSampleRate = 0.4;
+      options.replay.onErrorSampleRate = 1.0;
+    }, appRunner: () => runApp(SentryWidget(child: app)));
+  } else {
+    runApp(app);
+  }
 }
 
 class MoliseIsApp extends StatelessWidget {
@@ -56,9 +68,7 @@ class MoliseIsApp extends StatelessWidget {
     return MultiProvider(
       providers: providers,
       child: Consumer<ThemeViewModel>(
-        builder: (_, controller, _) {
-          _initializeSystemChrome(controller.brightness);
-
+        builder: (_, viewModel, _) {
           return MaterialApp.router(
             routerConfig: appRouter,
             builder: (context, child) {
@@ -88,7 +98,7 @@ class MoliseIsApp extends StatelessWidget {
             ],
             theme: AppThemeData.light(),
             darkTheme: AppThemeData.dark(),
-            themeMode: controller.themeMode,
+            themeMode: viewModel.themeMode,
             debugShowCheckedModeBanner: false,
           );
         },
