@@ -1,14 +1,19 @@
 abstract class RouteNames {
+  static const events = 'events';
+  static const eventsCategory = 'eventsCategory';
+  static const eventsCategoryDetails = 'eventsCategoryDetails';
+  static const eventsDetails = 'eventsDetails';
   static const favourites = 'favourites';
   static const favouritesCategory = 'favouritesCategory';
-  static const favouritesCategoryStory = 'favouritesCategoryStory';
+  static const favouritesCategoryDetail = 'favouritesCategoryStory';
   static const favouritesStory = 'favouritesStory';
   static const gallery = 'gallery';
   static const geoMap = 'geoMap';
   static const home = 'home';
   static const homeCategory = 'homeCategory';
-  static const homeCategoryStory = 'homeCategoryStory';
-  static const homeSearchResults = 'homeResults';
+  static const homeCategoryDetail = 'homeCategoryStory';
+  static const homeSearchResults = 'homeSearchResults';
+  static const homeSearchResultsDetail = 'homeSearchResultsStory';
   static const homeStory = 'homeStory';
   static const settings = 'settings';
   static const suggestion = 'suggestion';
