@@ -1,12 +1,13 @@
 abstract class RoutePaths {
   static const category = 'category/:index';
+  static const events = '/events';
   static const favourites = '/favourites';
   static const gallery = '/gallery';
   static const geoMap = '/map';
   static const home = '/';
-  static const homeSearchResults = 'results/:query';
+  static const homeSearchResults = 'search_results/:query';
   static const settings = '/settings';
   static const sync = '/sync';
-  static const story = 'story/:id';
+  static const details = 'details/:id';
   static const suggestion = 'suggestion';
 }
