@@ -1,44 +1,47 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
-import 'package:moliseis/data/repositories/core/repository_sync_result.dart';
-import 'package:moliseis/data/repositories/core/repository_sync_state.dart';
-import 'package:moliseis/domain/models/geo_map/geo_map_state.dart';
+import 'package:moliseis/domain/models/core/content_base.dart';
+import 'package:moliseis/domain/use-cases/explore/explore_use_case.dart';
+import 'package:moliseis/domain/use-cases/geo_map/geo_map_use_case.dart';
 import 'package:moliseis/routing/core_routes.dart';
 import 'package:moliseis/routing/route_names.dart';
 import 'package:moliseis/routing/route_paths.dart';
 import 'package:moliseis/ui/core/ui/scaffold_shell.dart';
+import 'package:moliseis/ui/event/view_models/event_view_model.dart';
+import 'package:moliseis/ui/event/widgets/events_screen.dart';
+import 'package:moliseis/ui/explore/view_models/explore_view_model.dart';
 import 'package:moliseis/ui/explore/widgets/explore_screen.dart';
 import 'package:moliseis/ui/favourite/view_models/favourite_view_model.dart';
 import 'package:moliseis/ui/favourite/widgets/favourite_screen.dart';
-import 'package:moliseis/ui/gallery/widgets/gallery_screen.dart';
 import 'package:moliseis/ui/geo_map/view_models/geo_map_view_model.dart';
 import 'package:moliseis/ui/geo_map/widgets/geo_map_screen.dart';
 import 'package:moliseis/ui/search/view_models/search_view_model.dart';
-import 'package:moliseis/ui/search/widgets/search_result.dart';
+import 'package:moliseis/ui/search/widgets/search_result_screen.dart';
 import 'package:moliseis/ui/settings/widgets/settings_screen.dart';
 import 'package:moliseis/ui/suggestion/view_models/suggestion_view_model.dart';
 import 'package:moliseis/ui/suggestion/widgets/suggestion_screen.dart';
 import 'package:moliseis/ui/sync/view_models/sync_view_model.dart';
 import 'package:moliseis/ui/sync/widgets/sync_screen.dart';
 import 'package:provider/provider.dart';
+import 'package:provider/single_child_widget.dart';
 
-final _rootNavigatorKey = GlobalKey<NavigatorState>();
+final _eventsShellNavigatorKey = GlobalKey<NavigatorState>();
 final _exploreShellNavigatorKey = GlobalKey<NavigatorState>();
-final _searchShellNavigatorKey = GlobalKey<NavigatorState>();
-final _galleryShellNavigatorKey = GlobalKey<NavigatorState>();
+final _favouritesShellNavigatorKey = GlobalKey<NavigatorState>();
 final _mapShellNavigatorKey = GlobalKey<NavigatorState>();
+final _rootNavigatorKey = GlobalKey<NavigatorState>();
 
 final appRouter = GoRouter(
   navigatorKey: _rootNavigatorKey,
   redirect: (context, state) {
-    final repositoryViewModel = context.read<SyncViewModel>();
+    final syncViewModel = context.read<SyncViewModel>();
 
-    if (repositoryViewModel.state == RepositorySyncState.loading) {
+    if (syncViewModel.sync.running) {
       return RoutePaths.sync;
     }
 
-    if (repositoryViewModel.state == RepositorySyncState.done &&
-        repositoryViewModel.result != RepositorySyncResult.majorError &&
+    if ((syncViewModel.sync.completed ||
+            syncViewModel.sync.error && !syncViewModel.fatalError) &&
         state.uri.toString().contains(RoutePaths.sync)) {
       return RoutePaths.home;
     }
@@ -66,29 +69,80 @@ final appRouter = GoRouter(
             GoRoute(
               path: RoutePaths.home,
               name: RouteNames.home,
-              builder: (_, _) => const ExploreScreen(),
+              builder: (_, _) {
+                return MultiProvider(
+                  providers: <SingleChildWidget>[
+                    ChangeNotifierProvider<EventViewModel>(
+                      create: (context) =>
+                          EventViewModel(repository: context.read())
+                            ..loadNextIds.execute(),
+                    ),
+                    ChangeNotifierProvider<ExploreViewModel>(
+                      create: (context) => ExploreViewModel(
+                        byIdUseCase: ExploreUseCase(
+                          eventRepository: context.read(),
+                          placeRepository: context.read(),
+                        ),
+                        placeRepository: context.read(),
+                      ),
+                    ),
+                    ChangeNotifierProvider<SearchViewModel>(
+                      create: (context) => SearchViewModel(
+                        eventRepository: context.read(),
+                        exploreGetByIdUseCase: ExploreUseCase(
+                          eventRepository: context.read(),
+                          placeRepository: context.read(),
+                        ),
+                        searchRepository: context.read(),
+                      ),
+                    ),
+                  ],
+                  builder: (context, _) {
+                    return ExploreScreen(
+                      eventViewModel: context.read(),
+                      exploreViewModel: context.read(),
+                      searchViewModel: context.read(),
+                    );
+                  },
+                );
+              },
               routes: <RouteBase>[
                 GoRoute(
                   path: RoutePaths.homeSearchResults,
                   name: RouteNames.homeSearchResults,
-                  builder: (context, state) {
-                    final viewModel = SearchViewModel(
-                      attractionRepository: context.read(),
-                      searchRepository: context.read(),
-                    );
-
+                  builder: (_, state) {
                     final query = state.pathParameters['query'] ?? '';
 
-                    viewModel.loadResults.execute(query);
-                    viewModel.loadRelatedResults.execute(query);
+                    return ChangeNotifierProvider(
+                      create: (context) {
+                        final viewModel = SearchViewModel(
+                          eventRepository: context.read(),
+                          exploreGetByIdUseCase: ExploreUseCase(
+                            eventRepository: context.read(),
+                            placeRepository: context.read(),
+                          ),
+                          searchRepository: context.read(),
+                        );
 
-                    return SearchResult(viewModel: viewModel, query: query);
+                        viewModel.loadResults.execute(query);
+                        // viewModel.loadRelatedResultsIds.execute(query);
+
+                        return viewModel;
+                      },
+                      builder: (context, _) => SearchResultScreen(
+                        query: query,
+                        viewModel: context.read(),
+                      ),
+                    );
                   },
+                  routes: <RouteBase>[
+                    detailRoute(name: RouteNames.homeSearchResultsDetail),
+                  ],
                 ),
-                storyRoute(name: RouteNames.homeStory),
-                categoriesRoute(
+                detailRoute(name: RouteNames.homeStory),
+                categoryRoute(
                   name: RouteNames.homeCategory,
-                  childName: RouteNames.homeCategoryStory,
+                  childName: RouteNames.homeCategoryDetail,
                 ),
                 GoRoute(
                   path: RoutePaths.suggestion,
@@ -106,7 +160,7 @@ final appRouter = GoRouter(
           ],
         ),
         StatefulShellBranch(
-          navigatorKey: _searchShellNavigatorKey,
+          navigatorKey: _favouritesShellNavigatorKey,
           routes: [
             GoRoute(
               path: RoutePaths.favourites,
@@ -117,24 +171,34 @@ final appRouter = GoRouter(
                 );
               },
               routes: <RouteBase>[
-                storyRoute(name: RouteNames.favouritesStory),
-                categoriesRoute(
+                detailRoute(name: RouteNames.favouritesStory),
+                categoryRoute(
                   name: RouteNames.favouritesCategory,
-                  childName: RouteNames.favouritesCategoryStory,
+                  childName: RouteNames.favouritesCategoryDetail,
                 ),
               ],
             ),
           ],
         ),
         StatefulShellBranch(
-          navigatorKey: _galleryShellNavigatorKey,
+          navigatorKey: _eventsShellNavigatorKey,
           routes: <RouteBase>[
             GoRoute(
-              path: RoutePaths.gallery,
-              name: RouteNames.gallery,
-              builder: (_, _) {
-                return const GalleryScreen();
+              path: RoutePaths.events,
+              name: RouteNames.events,
+              builder: (context, _) {
+                return ChangeNotifierProvider<EventViewModel>(
+                  create: (context) {
+                    return EventViewModel(repository: context.read());
+                  },
+                  child: Consumer<EventViewModel>(
+                    builder: (_, viewModel, _) {
+                      return EventsScreen(viewModel: viewModel);
+                    },
+                  ),
+                );
               },
+              routes: <RouteBase>[detailRoute(name: RouteNames.eventsDetails)],
             ),
           ],
         ),
@@ -145,13 +209,33 @@ final appRouter = GoRouter(
               path: RoutePaths.geoMap,
               name: RouteNames.geoMap,
               builder: (context, state) {
+                /*
                 final mapState =
-                    state.extra as GeoMapState? ?? const GeoMapState();
+                    state.extra as GeoMapState? ?? const GeoMapState(); */
+                final contentExtra = state.extra as ContentBase?;
+
                 final viewModel = GeoMapViewModel(
-                  attractionRepository: context.read(),
+                  geoMapUseCase: GeoMapUseCase(
+                    eventRepository: context.read(),
+                    geoMapRepository: context.read(),
+                    placeRepository: context.read(),
+                  ),
+                );
+                final searchViewModel = SearchViewModel(
+                  eventRepository: context.read(),
+                  exploreGetByIdUseCase: ExploreUseCase(
+                    eventRepository: context.read(),
+                    placeRepository: context.read(),
+                  ),
+                  searchRepository: context.read(),
                 );
 
-                return GeoMapScreen(mapState: mapState, viewModel: viewModel);
+                return GeoMapScreen(
+                  // mapState: mapState,
+                  contentExtra: contentExtra,
+                  viewModel: viewModel,
+                  searchViewModel: searchViewModel,
+                );
               },
             ),
           ],
