@@ -1,34 +1,43 @@
 import 'package:flutter/material.dart';
 import 'package:go_router/go_router.dart';
+import 'package:moliseis/domain/models/core/content_base.dart';
+import 'package:moliseis/domain/models/event/event_content.dart';
 import 'package:moliseis/routing/route_names.dart';
-import 'package:moliseis/ui/categories/widgets/category_button.dart';
-import 'package:moliseis/ui/core/ui/attraction_list_view_responsive.dart';
+import 'package:moliseis/ui/category/widgets/category_button.dart';
+import 'package:moliseis/ui/core/ui/content/content_adaptive_list_grid_view.dart';
+import 'package:moliseis/ui/core/ui/empty_view.dart';
+import 'package:moliseis/ui/core/ui/skeletons/skeleton_content_grid.dart';
+import 'package:moliseis/ui/core/ui/skeletons/skeleton_content_list.dart';
 import 'package:moliseis/ui/core/ui/text_section_divider.dart';
-import 'package:moliseis/ui/explore/view_models/attraction_view_model.dart';
+import 'package:moliseis/ui/event/view_models/event_view_model.dart';
+import 'package:moliseis/ui/explore/view_models/explore_view_model.dart';
 import 'package:moliseis/ui/explore/widgets/explore_screen_carousel_view.dart';
+import 'package:moliseis/ui/search/view_models/search_view_model.dart';
 import 'package:moliseis/ui/search/widgets/custom_search_anchor.dart';
 import 'package:moliseis/ui/suggestion/widgets/suggestion_cta_button.dart';
 import 'package:moliseis/ui/sync/view_models/sync_view_model.dart';
 import 'package:moliseis/utils/constants.dart';
 import 'package:provider/provider.dart';
+import 'package:responsive_framework/responsive_framework.dart';
 
 class ExploreScreen extends StatefulWidget {
-  const ExploreScreen({super.key});
+  const ExploreScreen({
+    super.key,
+    required this.eventViewModel,
+    required this.exploreViewModel,
+    required this.searchViewModel,
+  });
+
+  final EventViewModel eventViewModel;
+  final ExploreViewModel exploreViewModel;
+  final SearchViewModel searchViewModel;
 
   @override
   State<ExploreScreen> createState() => _ExploreScreenState();
 }
 
 class _ExploreScreenState extends State<ExploreScreen> {
-  late final AttractionViewModel _viewModel;
   final _searchController = SearchController();
-
-  @override
-  void initState() {
-    super.initState();
-
-    _viewModel = context.read();
-  }
 
   @override
   void dispose() {
@@ -43,7 +52,7 @@ class _ExploreScreenState extends State<ExploreScreen> {
         child: RefreshIndicator(
           edgeOffset: kToolbarHeight * 2.0 + 8.0,
           onRefresh: () async {
-            context.read<SyncViewModel>().synchronize(force: true);
+            context.read<SyncViewModel>().sync.execute(true);
 
             // Redirects to the local app repositories synchronization screen.
             GoRouter.of(context).refresh();
@@ -81,8 +90,8 @@ class _ExploreScreenState extends State<ExploreScreen> {
                 ],
               ),
               SliverAppBar(
-                elevation: 0.0,
-                scrolledUnderElevation: 0.0,
+                elevation: 0,
+                scrolledUnderElevation: 0,
                 backgroundColor: Theme.of(context).colorScheme.surface,
                 flexibleSpace: Align(
                   alignment: Alignment.centerLeft,
@@ -91,10 +100,11 @@ class _ExploreScreenState extends State<ExploreScreen> {
                     onSubmitted: (text) {
                       _showSearchResults(text);
                     },
-                    onSuggestionPressed: (_) {
+                    onSuggestionPressed: (ContentBase content) {
                       _searchController.closeView(null);
-                      _showSearchResults(_searchController.text);
+                      _showSearchResults(content.name);
                     },
+                    viewModel: widget.searchViewModel,
                   ),
                 ),
                 primary: false,
@@ -105,7 +115,7 @@ class _ExploreScreenState extends State<ExploreScreen> {
               ),
               const SliverToBoxAdapter(child: SizedBox(height: 8.0)),
               ExploreScreenCarouselView(
-                attractionsIdsFuture: _viewModel.suggestedAttractionIds,
+                exploreViewModel: widget.exploreViewModel,
               ),
               const SliverPadding(
                 padding: EdgeInsetsDirectional.fromSTEB(16.0, 16.0, 16.0, 8.0),
@@ -119,19 +129,19 @@ class _ExploreScreenState extends State<ExploreScreen> {
                 ),
                 sliver: SliverGrid(
                   delegate: SliverChildBuilderDelegate((context, index) {
-                    final attractionType = _viewModel.attractionTypes[index];
+                    final type = widget.exploreViewModel.types[index];
                     return CategoryButton(
                       onPressed: () {
                         return GoRouter.of(context).goNamed(
                           RouteNames.homeCategory,
                           pathParameters: {
-                            'index': (attractionType.index - 1).toString(),
+                            'index': (type.index - 1).toString(),
                           },
                         );
                       },
-                      attractionType: attractionType,
+                      contentCategory: type,
                     );
-                  }, childCount: _viewModel.attractionTypes.length),
+                  }, childCount: widget.exploreViewModel.types.length),
                   gridDelegate: const SliverGridDelegateWithFixedCrossAxisCount(
                     crossAxisCount: 2,
                     mainAxisSpacing: 8.0,
@@ -148,11 +158,67 @@ class _ExploreScreenState extends State<ExploreScreen> {
                   ),
                 ),
               ),
+              const SliverPadding(
+                padding: EdgeInsets.symmetric(horizontal: 16.0),
+                sliver: SliverToBoxAdapter(
+                  child: TextSectionDivider('Prossimi eventi'),
+                ),
+              ),
+              ListenableBuilder(
+                listenable: widget.eventViewModel.loadNext,
+                builder: (context, child) {
+                  if (widget.eventViewModel.loadNext.completed) {
+                    return SliverPadding(
+                      padding: const EdgeInsets.only(top: 8.0, bottom: 16.0),
+                      sliver: ContentAdaptiveListGridView(
+                        widget.eventViewModel.next,
+                        onPressed: (content) {
+                          GoRouter.of(context).goNamed(
+                            RouteNames.homeStory,
+                            pathParameters: {'id': content.remoteId.toString()},
+                            queryParameters: {
+                              'isEvent': (content is EventContent
+                                  ? 'true'
+                                  : 'false'),
+                            },
+                          );
+                        },
+                      ),
+                    );
+                  }
+
+                  if (widget.eventViewModel.loadNext.error) {
+                    return SliverPadding(
+                      padding: const EdgeInsets.fromLTRB(16.0, 0, 16.0, 16.0),
+                      sliver: SliverToBoxAdapter(
+                        child: EmptyView.error(
+                          text: const Text(
+                            'Si è verificato un errore durante il caricamento.',
+                          ),
+                          action: TextButton(
+                            onPressed: () {
+                              widget.eventViewModel.loadNext.execute();
+                            },
+                            child: const Text('Riprova'),
+                          ),
+                        ),
+                      ),
+                    );
+                  }
+
+                  final length = widget.eventViewModel.nextIds.length;
+
+                  return ResponsiveBreakpoints.of(context).isMobile
+                      ? SkeletonContentList.sliver(itemCount: length)
+                      : CardSkeletonGrid.sliver(itemCount: length);
+                },
+              ),
               SliverPadding(
                 padding: const EdgeInsets.symmetric(horizontal: 16.0),
                 sliver: SliverToBoxAdapter(
                   child: Row(
                     mainAxisAlignment: MainAxisAlignment.spaceBetween,
+                    spacing: 16.0,
                     children: [
                       const Expanded(
                         child: TextSectionDivider('Ultimi aggiunti'),
@@ -160,7 +226,9 @@ class _ExploreScreenState extends State<ExploreScreen> {
                       OutlinedButton.icon(
                         onPressed: () => GoRouter.of(context).goNamed(
                           RouteNames.homeCategory,
-                          pathParameters: {'index': '0'},
+                          pathParameters: {
+                            'index': kcategoryScreenNoIndex.toString(),
+                          },
                         ),
                         style: const ButtonStyle(
                           visualDensity: VisualDensity.compact,
@@ -173,13 +241,50 @@ class _ExploreScreenState extends State<ExploreScreen> {
                 ),
               ),
               const SliverToBoxAdapter(child: SizedBox(height: 8.0)),
-              AttractionListViewResponsive(
-                _viewModel.latestAttractionIds,
-                onPressed: (attractionId) {
-                  GoRouter.of(context).goNamed(
-                    RouteNames.homeStory,
-                    pathParameters: {'id': attractionId.toString()},
-                  );
+              ListenableBuilder(
+                listenable: widget.exploreViewModel.loadLatest,
+                builder: (context, child) {
+                  if (widget.exploreViewModel.loadLatest.completed) {
+                    return ContentAdaptiveListGridView(
+                      widget.exploreViewModel.latest,
+                      onPressed: (content) {
+                        GoRouter.of(context).goNamed(
+                          RouteNames.homeStory,
+                          pathParameters: {'id': content.remoteId.toString()},
+                          queryParameters: {
+                            'isEvent': (content is EventContent
+                                ? 'true'
+                                : 'false'),
+                          },
+                        );
+                      },
+                    );
+                  }
+
+                  if (widget.exploreViewModel.loadLatest.error) {
+                    return SliverPadding(
+                      padding: const EdgeInsets.fromLTRB(16.0, 0, 16.0, 16.0),
+                      sliver: SliverToBoxAdapter(
+                        child: EmptyView.error(
+                          text: const Text(
+                            'Si è verificato un errore durante il caricamento.',
+                          ),
+                          action: TextButton(
+                            onPressed: () {
+                              widget.exploreViewModel.loadLatest.execute();
+                            },
+                            child: const Text('Riprova'),
+                          ),
+                        ),
+                      ),
+                    );
+                  }
+
+                  final length = widget.exploreViewModel.latestIds.length;
+
+                  return ResponsiveBreakpoints.of(context).isMobile
+                      ? SkeletonContentList.sliver(itemCount: length)
+                      : CardSkeletonGrid.sliver(itemCount: length);
                 },
               ),
               const SliverToBoxAdapter(child: SizedBox(height: 16.0)),
