@@ -2,6 +2,7 @@ import 'package:flutter/material.dart';
 import 'package:moliseis/ui/core/ui/custom_snack_bar.dart';
 import 'package:moliseis/ui/core/ui/url_text_button.dart';
 import 'package:moliseis/utils/app_url_launcher.dart';
+import 'package:moliseis/utils/constants.dart';
 import 'package:provider/provider.dart';
 
 class GeoMapAttribution extends StatelessWidget {
@@ -9,50 +10,50 @@ class GeoMapAttribution extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    return Align(
-      alignment: Alignment.bottomRight,
-      child: Wrap(
-        alignment: WrapAlignment.end,
-        spacing: 8.0,
-        runAlignment: WrapAlignment.end,
-        crossAxisAlignment: WrapCrossAlignment.center,
-        children: <Widget>[
-          UrlTextButton.icon(
-            onPressed: () async {
-              if (!await context.read<AppUrlLauncher>().mapTilerWebsite()) {
-                if (context.mounted) {
-                  showSnackBar(
-                    context: context,
-                    textContent:
-                        'Si è verificato un errore, riprova più tardi.',
-                  );
+    return Padding(
+      padding: const EdgeInsets.symmetric(horizontal: 16.0),
+      child: Align(
+        alignment: Alignment.bottomRight,
+        child: Wrap(
+          alignment: WrapAlignment.end,
+          spacing: 8.0,
+          runAlignment: WrapAlignment.end,
+          crossAxisAlignment: WrapCrossAlignment.center,
+          children: <Widget>[
+            UrlTextButton.icon(
+              onPressed: () async {
+                if (!await context.read<AppUrlLauncher>().mapTilerWebsite()) {
+                  if (context.mounted) {
+                    showSnackBar(
+                      context: context,
+                      textContent:
+                          'Si è verificato un errore, riprova più tardi.',
+                    );
+                  }
                 }
-              }
-            },
-            icon: const ImageIcon(
-              AssetImage('assets/icon/maptiler-icon-dark_64x69_optimized.png'),
+              },
+              icon: const ImageIcon(AssetImage(kAssetMapTilerIconPath)),
+              iconSize: 24.0,
+              label: const Text('© MapTiler'),
             ),
-            iconSize: 24.0,
-            label: const Text('© MapTiler'),
-          ),
-          UrlTextButton(
-            onPressed: () async {
-              if (!await context
-                  .read<AppUrlLauncher>()
-                  .openStreetMapWebsite()) {
-                if (context.mounted) {
-                  showSnackBar(
-                    context: context,
-                    textContent:
-                        'Si è verificato un errore, riprova più tardi.',
-                  );
+            UrlTextButton(
+              onPressed: () async {
+                if (!await context
+                    .read<AppUrlLauncher>()
+                    .openStreetMapWebsite()) {
+                  if (context.mounted) {
+                    showSnackBar(
+                      context: context,
+                      textContent:
+                          'Si è verificato un errore, riprova più tardi.',
+                    );
+                  }
                 }
-              }
-            },
-            label: const Text('© OpenStreetMap contributors'),
-          ),
-          const SizedBox(width: 8.0),
-        ],
+              },
+              label: const Text('© OpenStreetMap contributors'),
+            ),
+          ],
+        ),
       ),
     );
   }
