@@ -1,5 +1,6 @@
 import 'package:flutter/material.dart';
-import 'package:moliseis/ui/core/themes/text_style.dart';
+import 'package:moliseis/ui/core/themes/text_styles.dart';
+import 'package:moliseis/ui/core/ui/cards/card_base.dart';
 
 class SuggestionCTAButton extends StatelessWidget {
   const SuggestionCTAButton({super.key, required this.onPressed});
@@ -8,54 +9,44 @@ class SuggestionCTAButton extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final bgColor = Theme.of(context).colorScheme.tertiaryFixed;
-    final fgColor = Theme.of(context).colorScheme.onTertiaryFixed;
+    final bgColor = Theme.of(context).colorScheme.secondaryContainer;
+    final fgColor = Theme.of(context).colorScheme.onSecondaryContainer;
     return Column(
       mainAxisSize: MainAxisSize.min,
-      children: [
-        ConstrainedBox(
-          constraints: const BoxConstraints(minHeight: 112.0),
-          child: Card(
-            color: bgColor,
-            elevation: 0.0,
-            shape: RoundedRectangleBorder(
-              borderRadius: BorderRadius.circular(26.0),
-            ),
-            margin: EdgeInsets.zero,
-            clipBehavior: Clip.hardEdge,
-            child: InkWell(
-              onTap: onPressed,
-              child: Padding(
-                padding: const EdgeInsets.all(16.0),
-                child: Row(
-                  crossAxisAlignment: CrossAxisAlignment.start,
-                  children: [
-                    Expanded(
-                      child: Padding(
-                        padding: const EdgeInsetsDirectional.only(end: 16.0),
-                        child: Column(
-                          crossAxisAlignment: CrossAxisAlignment.start,
-                          spacing: 4.0,
-                          children: [
-                            Text(
-                              'Suggerisci un luogo o evento',
-                              style: CustomTextStyles.titleSmaller(
-                                context,
-                              )?.copyWith(color: fgColor),
-                            ),
-                            Text(
-                              'Fai scoprire i migliori luoghi o eventi che il '
-                              'Molise ha da offrire!',
-                              style: TextStyle(color: fgColor),
-                            ),
-                          ],
+      children: <Widget>[
+        CardBase(
+          color: bgColor,
+          elevation: 0,
+          onPressed: onPressed,
+          child: Padding(
+            padding: const EdgeInsets.all(16.0),
+            child: Row(
+              crossAxisAlignment: CrossAxisAlignment.start,
+              children: <Widget>[
+                Expanded(
+                  child: Padding(
+                    padding: const EdgeInsetsDirectional.only(end: 16.0),
+                    child: Column(
+                      crossAxisAlignment: CrossAxisAlignment.start,
+                      spacing: 4.0,
+                      children: <Widget>[
+                        Text(
+                          'Suggerisci un luogo o un evento',
+                          style: CustomTextStyles.titleSmaller(
+                            context,
+                          )?.copyWith(color: fgColor),
                         ),
-                      ),
+                        Text(
+                          'Fai scoprire i migliori luoghi o eventi che il '
+                          'Molise ha da offrire!',
+                          style: TextStyle(color: fgColor),
+                        ),
+                      ],
                     ),
-                    Icon(Icons.maps_ugc, color: fgColor),
-                  ],
+                  ),
                 ),
-              ),
+                Icon(Icons.maps_ugc, color: fgColor),
+              ],
             ),
           ),
         ),
