@@ -1,6 +1,8 @@
+const double kGridViewCardWidth = 252.0;
+
 const double kGridViewCardHeight = 168.0;
 
-const double kEventCardHeight = 240.0;
+const double kEventGridViewCardHeight = 200.0;
 
 const double kButtonHeight = 40.0;
 
@@ -13,3 +15,8 @@ const double kStorySlideshowLandscapeHeightPerc = 0.65;
 const double kStorySlideshowPortraitPainterHeightPerc = 0.13;
 
 const double kStorySlideshowLandscapePainterHeightPerc = 0.195;
+
+const int kcategoryScreenNoIndex = -1;
+
+const String kAssetMapTilerIconPath =
+    'assets/icons/maptiler-icon-dark_64x69_optimized.png';
