@@ -1,43 +1,11 @@
-class AttractionNullException implements Exception {
-  const AttractionNullException(this.id);
+class CityNullException implements Exception {
+  const CityNullException(this.id);
 
   final int id;
 
   @override
   String toString() {
-    return 'No Attraction with id: $id has been found in the Attractions box.';
-  }
-}
-
-class MolisImageNullException implements Exception {
-  const MolisImageNullException(this.id);
-
-  final int id;
-
-  @override
-  String toString() {
-    return 'No MolisImage from attraction id: $id has been found in the '
-        'MolisImages box.';
-  }
-}
-
-class PlaceNullException implements Exception {
-  const PlaceNullException(this.id);
-
-  final int id;
-
-  @override
-  String toString() {
-    return 'No Place from attraction id: $id has been found in the Places box.';
-  }
-}
-
-class SupabaseResponseException implements Exception {
-  const SupabaseResponseException();
-
-  @override
-  String toString() {
-    return 'Supabase exception.';
+    return 'No City from Place remote ID: $id has been found in the City box.';
   }
 }
 
@@ -65,5 +33,27 @@ class CloudinaryNullResponseException implements Exception {
   @override
   String toString() {
     return 'Cloudinary returned no response (null).';
+  }
+}
+
+class MediaNullException implements Exception {
+  const MediaNullException(this.id);
+
+  final int id;
+
+  @override
+  String toString() {
+    return 'No Media from Place remote ID: $id has been found in the Media box.';
+  }
+}
+
+class PlaceNullException implements Exception {
+  const PlaceNullException(this.id);
+
+  final int id;
+
+  @override
+  String toString() {
+    return 'No Place with remote ID: $id has been found in the Places box.';
   }
 }
