@@ -1,42 +1,50 @@
 import 'package:flutter/material.dart';
-import 'package:moliseis/domain/models/attraction/attraction_type.dart';
+import 'package:moliseis/domain/models/core/content_category.dart';
+import 'package:moliseis/domain/models/core/content_type.dart';
 
-extension AttractionTypeExtensions on AttractionType {
+extension ContentCategoryExtensions on ContentCategory {
   IconData get icon => switch (this) {
-    AttractionType.unknown => Icons.question_mark,
-    AttractionType.nature => Icons.forest,
-    AttractionType.history => Icons.school,
-    AttractionType.folklore => Icons.groups_3,
-    AttractionType.food => Icons.fastfood,
-    AttractionType.allure => Icons.attractions,
-    AttractionType.experience => Icons.signpost,
+    ContentCategory.unknown => Icons.question_mark,
+    ContentCategory.nature => Icons.forest,
+    ContentCategory.history => Icons.school,
+    ContentCategory.folklore => Icons.groups_3,
+    ContentCategory.food => Icons.fastfood,
+    ContentCategory.allure => Icons.attractions,
+    ContentCategory.experience => Icons.signpost,
   };
 
   IconData get iconAlt => switch (this) {
-    AttractionType.unknown => Icons.question_mark_outlined,
-    AttractionType.nature => Icons.forest_outlined,
-    AttractionType.history => Icons.school_outlined,
-    AttractionType.folklore => Icons.groups_3_outlined,
-    AttractionType.food => Icons.fastfood_outlined,
-    AttractionType.allure => Icons.attractions_outlined,
-    AttractionType.experience => Icons.signpost_outlined,
+    ContentCategory.unknown => Icons.question_mark_outlined,
+    ContentCategory.nature => Icons.forest_outlined,
+    ContentCategory.history => Icons.school_outlined,
+    ContentCategory.folklore => Icons.groups_3_outlined,
+    ContentCategory.food => Icons.fastfood_outlined,
+    ContentCategory.allure => Icons.attractions_outlined,
+    ContentCategory.experience => Icons.signpost_outlined,
   };
 
   String get label => switch (this) {
-    AttractionType.unknown => 'unknown',
-    AttractionType.nature => 'Natura',
-    AttractionType.history => 'Storia',
-    AttractionType.folklore => 'Folklore',
-    AttractionType.food => 'Cibo',
-    AttractionType.allure => 'Attrazioni',
-    AttractionType.experience => 'Esperienze',
+    ContentCategory.unknown => 'unknown',
+    ContentCategory.nature => 'Natura',
+    ContentCategory.history => 'Storia',
+    ContentCategory.folklore => 'Folklore',
+    ContentCategory.food => 'Cibo',
+    ContentCategory.allure => 'Attrazioni',
+    ContentCategory.experience => 'Esperienze',
   };
 }
 
-extension AttractionTypeListExtensions on List<AttractionType> {
-  /// Returns a list of [AttractionType]s containing all but
-  /// [AttractionType.unknown].
-  List<AttractionType> get minusUnknown => AttractionType.values.sublist(1);
+extension ContentTypeExtensions on ContentType {
+  String get label => switch (this) {
+    ContentType.event => 'Eventi',
+    ContentType.place => 'Luoghi',
+  };
+}
+
+extension ContentCategoryListExtensions on List<ContentCategory> {
+  /// Returns a list of [ContentCategory]s containing all but
+  /// [ContentCategory.unknown].
+  List<ContentCategory> get minusUnknown => ContentCategory.values.sublist(1);
 }
 
 extension BuildContextExtensions on BuildContext {
@@ -51,4 +59,6 @@ extension BuildContextExtensions on BuildContext {
   bool get isDarkTheme {
     return Theme.of(this).brightness == Brightness.dark;
   }
+
+  // TODO(xmattjus): Implement a method to check if the device is a phone.
 }
